(svn r2080) - Fix: [OldAI] p->ai.state_counter was uint16 but could hold a tile index, therefore overflowing - changed that to uint32 and bumped the savegame revision to 32. It *MIGHT* close bug 1151374 - it certainly caused AI to stop building anything sometimes. - While at it, use TileIndex as the tile index type in AiRemovePlayerRailOrRoad() and AiStateRemoveTrack(). - Make the number of tiles scanned 4*MapSizeX() instead of 1000. It *MIGHT* close bug 1116614.

+++ player.h
@@ -27,7 +27,7 @@ typedef struct AiBuildRec {
 typedef struct PlayerAI {
 	byte state;
 	byte tick; // Used to determine how often to move
-	uint16 state_counter;
+	uint32 state_counter; // Can hold tile index!
 	uint16 timeout_counter;
 
 	byte state_mode;
